# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Couvoit API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "List brands",
                "description": "List brands.",
                "operationId": "brandsIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/brands/{id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get brand by id",
                "description": "Show a brand.",
                "operationId": "brandsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "List cars",
                "description": "List cars.\nAdmin: list all cars\nUser: only his car",
                "operationId": "carsIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create my car",
                "description": "Create a car for the authenticated user.",
                "operationId": "carsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCarRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get car by id",
                "description": "Show a car.",
                "operationId": "carsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete my car",
                "description": "Delete a car.",
                "operationId": "carsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Cars"
                ],
                "summary": "Update my car",
                "description": "Update a car.",
                "operationId": "carsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCarRequestPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCarRequestPayload": {
                "required": [
                    "brand",
                    "type",
                    "model",
                    "color",
                    "carregistration"
                ],
                "properties": {
                    "brand": {
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "peugeot"
                            }
                        },
                        "type": "object"
                    },
                    "type": {
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "berline"
                            }
                        },
                        "type": "object"
                    },
                    "model": {
                        "required": [
                            "name",
                            "seats"
                        ],
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "308"
                            },
                            "seats": {
                                "type": "integer",
                                "maximum": 9,
                                "minimum": 1,
                                "example": 5
                            }
                        },
                        "type": "object"
                    },
                    "color": {
                        "required": [
                            "name",
                            "hex_code"
                        ],
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "bleu"
                            },
                            "hex_code": {
                                "type": "string",
                                "example": "#00AAFF"
                            }
                        },
                        "type": "object"
                    },
                    "carregistration": {
                        "type": "string",
                        "example": "AA123BB"
                    },
                    "seats": {
                        "type": "integer",
                        "maximum": 9,
                        "minimum": 1,
                        "example": 5,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateCarRequestPayload": {
                "properties": {
                    "carregistration": {
                        "type": "string",
                        "example": "AA123BB",
                        "nullable": true
                    },
                    "license_plate": {
                        "type": "string",
                        "example": "AA123BB",
                        "nullable": true
                    },
                    "color": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "bleu",
                                "nullable": true
                            },
                            "hex_code": {
                                "type": "string",
                                "example": "#00AAFF",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "model": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "308",
                                "nullable": true
                            },
                            "seats": {
                                "type": "integer",
                                "maximum": 9,
                                "minimum": 1,
                                "example": 5,
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "brand": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "peugeot",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "type": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "berline",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}